+++ Tempest/2d/surfacerender.h
@@ -140,9 +140,9 @@ class SurfaceRender {
 
       virtual void quad( int x, int y, int  w, int  h,
                          int tx, int ty, int tw, int th );
-      virtual void triangle( int x1, int y1, int u1, int v1,
-                             int x2, int y2, int u2, int v2,
-                             int x3, int y3, int u3, int v3 );
+      virtual void triangle( float x1, float y1, float u1, float v1,
+                             float x2, float y2, float u2, float v2,
+                             float x3, float y3, float u3, float v3 );
       virtual void line( int x, int y, int x2, int y2);
 
       virtual void setBlendMode( BlendMode m );
@@ -166,9 +166,9 @@ class SurfaceRender {
 
     virtual void quad( int x, int y, int  w, int  h,
                        int tx, int ty, int tw, int th );
-    virtual void triangle( int x1, int y1, int u1, int v1,
-                           int x2, int y2, int u2, int v2,
-                           int x3, int y3, int u3, int v3 );
+    virtual void triangle( float x1, float y1, float u1, float v1,
+                           float x2, float y2, float u2, float v2,
+                           float x3, float y3, float u3, float v3 );
     virtual void line( int x, int y, int x2, int y2);
 
     const Tempest::VertexDeclaration::Declarator &decl();
